extract the storage module
