+++ adapter/pom.xml
@@ -26,6 +26,7 @@
         <module>api</module>
         <module>rpc</module>
         <module>runtimer</module>
+        <module>storage</module>
     </modules>
 
 </project>
